requirement.txt
argcomplete==1.10.0
attrs==19.3.0
beautifulsoup4==4.8.0
blis==0.4.1
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.3
Django==2.2.8
django-hosts==3.0
django-rest-swagger==2.2.0
django-seed==0.1.9
django-upload-validator==1.1.5
djangorestframework==3.10.3
docx2txt==0.8
drf-yasg==1.17.0
EbookLib==0.17.1
en-core-web-sm==2.2.5
extract-msg==0.23.1
Faker==3.0.0
idna==2.8
IMAPClient==2.1.0
importlib-metadata==1.3.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.0.2
murmurhash==1.0.2
newsapi==0.1.1
newsapi-python==0.2.6
nltk==3.4.5
numpy==1.17.4
olefile==0.46
openapi-codec==1.3.2
packaging==19.2
pandas==0.24.2
pdfminer.six==20181108
Pillow==6.2.1
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
pycryptodome==3.9.4
pyparsing==2.4.5
PyPDF2==1.26.0
pyresparser==1.0.6
pyrsistent==0.15.6
python-dateutil==2.8.1
python-magic==0.4.15
python-pptx==0.6.18
pytz==2019.3
PyYAML==5.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.2.3
SpeechRecognition==3.8.1
sqlparse==0.3.0
srsly==0.2.0
swagger-spec-validator==2.4.3
text-unidecode==1.3
textract==1.6.3
thinc==7.3.1
tqdm==4.40.2
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.25.7
wasabi==0.4.2
xlrd==1.2.0
XlsxWriter==1.2.6
zipp==0.6.0
